Search and replace every of occurence of course by group
Warning! In order to be able to use DocHub once this change is done, you have
to completely reset all database/cache stuff. The best I found for this is to
delete the whole project directory and git clone again from source.

If you attempt to import this commit when the site has already been built, there
probably be some strange runtime or test-time errors.