Move the new thread form on the thread list page
We can now post a new thread of dicussion straight from the thread list on a
given group. This allows faster usability (load a new page to ask a quesiton
is annoying) and a more Facebook-like feel (which is very important if we want
that users move on from Facebook).