Show the whole threads in group page
Now all the content of all thread is shown on the group page, and the message
form is also on this page.

For that, we have to import the message form, tweak a bit its template in order
to get rid of the box around it (this box is not essential, and pollute the
design in the group page), and effectively show all messages on the group
template.