Add chat app, and bind it to websocket messages
I created an app because I didn't know where to put the Message class,
as it isn't part of 'catalog' nor 'telepathy'. This also allows us to
put consumers.py in a more appropriate place.

We could even put the template of the chat in a 'chat' template
folder, and just include the template in 'group.html'