Change the way pad client handles user input and server response
Previously, when the user typed a key, the change on the pad was
calculated, sent to server and locally reverted. And only once the
server response came, the user change (validated by the server) could
be applied to the local textarea.

Now, the user input is not reverted, but rather buffered in a variable,
besides the change being sent to server. And when a message come from
the server, the server edit is compared to the buffered user edit. If
they match, then nothing is done. But is they don't match, then the
buffered user edit is reverted and the server edit is applied.

This behavior changes nothing on the server side, and requires less
live changes in the client textarea, which leads to less bugs and
glitches.